+++ src/Entity/FieldOption.php
@@ -1,0 +1,90 @@
+<?php
+
+/*
+ * This file is part of the API as a Service Project.
+ *
+ * Copyright (c) 2019 Christian Siewert <christian@sieware.international>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace App\Entity;
+
+use Doctrine\ORM\Mapping as ORM;
+use ApiPlatform\Core\Annotation\ApiResource;
+
+/**
+ * Key-value pairs of options that get passed to the underlying
+ * database platform when generating DDL statements.
+ *
+ * @ApiResource()
+ * @ORM\Entity()
+ * @author Christian Siewert <christian@sieware.international>
+ */
+class FieldOption
+{
+    /**
+     * @ORM\Id()
+     * @ORM\GeneratedValue()
+     * @ORM\Column(type="integer")
+     */
+    private $id;
+
+    /**
+     * @ORM\Column(type="string", length=255)
+     */
+    private $name;
+
+    /**
+     * @ORM\Column(type="string", length=255)
+     */
+    private $value;
+
+    /**
+     * @ORM\ManyToOne(targetEntity="App\Entity\ServiceField", inversedBy="options")
+     * @ORM\JoinColumn(nullable=false)
+     */
+    private $serviceField;
+
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    public function getName(): ?string
+    {
+        return $this->name;
+    }
+
+    public function setName(string $name): self
+    {
+        $this->name = $name;
+
+        return $this;
+    }
+
+    public function getValue(): ?string
+    {
+        return $this->value;
+    }
+
+    public function setValue(string $value): self
+    {
+        $this->value = $value;
+
+        return $this;
+    }
+
+    public function getServiceField(): ?ServiceField
+    {
+        return $this->serviceField;
+    }
+
+    public function setServiceField(?ServiceField $serviceField): self
+    {
+        $this->serviceField = $serviceField;
+
+        return $this;
+    }
+}